[QA-317] Update metadata checks (#456)
* update checks

* more updates

* split generic checks

* rename

* remove names

* update author name

* move proxy check

* format

* update comments, add todos

* add not all caps check

* combine journal ref checks

* genericize journal ref checks, refactor content checks

* another small refactor

* format

* make year check a pattern check

* update acm and msc

* update readme, bugfix required_inputs, update tests

* allow overrides of failure_message on generic checks

* update messages

* update author names pattern, messages

* update et al check, messages

* update valid doi check, messages

* update author name tests

* bugfix invert valid doi check

* move structure checks

* finish updating messages

* update more messages

* update doi checks

* update whitespace and abstract content checks

* override messages with more info

* remove duplicate journal ref length check, fix author length typo

* fix journal ref test

* revert authors length to 10000, remove comma case from whitespace check
