--- qa/qa/checks/generic/all_caps_words.py
@@ -1,801 +1,0 @@
-"""
-Words of 6 characters or more that are known to correctly appear in
-capitals and should thus be ignored by arXiv admin checks.
-
-Copied from package arXiv::Submit::MetaCheck_CAPS;
-"""
-
-known_words_in_all_caps_str = """AADEBUG
-AASLATEX
-AASTEX
-ABBOTT
-ABLAUT
-ABLHASAN
-ABRIDGED
-ACERBI
-ADONIS
-ADVANCES
-AGLIETTI
-AIHENP
-AIKAWA
-AIRAPT
-AKIMOV
-AKOROTKOV
-ALADIN
-ALAVAGNO
-ALBERICO
-ALBERTA
-ALFONSO
-ALISHAHIHA
-ALLFRAME
-AMANDA
-AMSLATEX
-AMSPPT
-AMSSYM
-AMSTEX
-ANAHIT
-ANAZAWA
-ANCHORDOQUI
-ANDERSEN
-ANDREA
-ANDREANI
-ANDRIANOPOLI
-ANDRIANOV
-ANGELA
-ANGELOW
-ANTARES
-ANTONOV
-ANTONUCCIO
-ANUATC
-ARANJAN
-ARISUE
-ARKANI
-ASCHIERI
-ATHENA
-AUGIER
-AUTOFIB
-AXODRAW
-BACODINE
-BAIKAL
-BALAZS
-BALLESTRERO
-BALQSO
-BANERJEE
-BARBERO
-BARBIERI
-BARONE
-BARRAT
-BASSETTO
-BAULIEU
-BCHIBISOV
-BECATTINI
-BECCHI
-BEHNIA
-BELLUCCI
-BERCEANU
-BERERA
-BEREZHIANI
-BERGLIAFFA
-BERING
-BERNARDO
-BERVILLIER
-BGEARN
-BHAGEN
-BHARMS
-BIANCHI
-BIFERALE
-BLANCA
-BLESSING
-BNUPHY
-BONATSOS
-BOOMER
-BOOMERANG
-BOREXINO
-BOSTED
-BRAHMS
-BRIGNOLE
-BUCHALLA
-BULGAC
-BUTURI
-BVIJHK
-BVRJHK
-CABALLERO
-CALVANI
-CAMICI
-CANGAROO
-CAPEZZALI
-CARACCIOLO
-CAROLINE
-CARONE
-CARRETTA
-CARRUTHERS
-CARTER
-CASCADE
-CASELLE
-CASPEC
-CASSINI
-CASSISI
-CASTOR
-CATELAN
-CDMTCS
-CEADRF
-CELEBONOVIC
-CELESTE
-CENTER
-CETDST
-CHAMATI
-CHANDRA
-CHAVES
-CHIEZE
-CHORUS
-CHOUDHURY
-CHRVLZH
-CICUTA
-CIMENTO
-CINVESTAV
-CIOTTI
-CIPANP
-CIRIGLIANO
-CITAMU
-CLEMENT
-CLOUDY
-CMBFAST
-CNELSON
-COBRAS
-COLDEA
-COLING
-COMPASS
-COMPTEL
-COMPTON
-CONFERENCE
-CONICET
-CONNIE
-CONTEXT
-CORAVEL
-CORBEL
-CORNELL
-CORSIKA
-COSMION
-COSMOS
-COSPAR
-CPLEAR
-CRENSHAW
-CSIC
-CTSTATEU
-CXXXIII
-CXXXIV
-CYCOMP
-DANIEL
-DAOPHOT
-DAPHNE
-DAPNIA
-DAPNIS
-DASWAM
-DAWSON
-DEBOER
-DECNET
-DELDUC
-DELEOS
-DELFINO
-DELPHI
-DEMARTEAU
-DEMIRM
-DESILVA
-DHURJATI
-DICKMAN
-DIQUARKS
-DIRECT
-DISASTER
-DLBBAT
-DOMINICI
-DONIZETI
-DORIGO
-DREINER
-DRYUMA
-DSYIBM
-DURHAM
-DYNAMIC
-EDINBURGH
-EETDST
-EFFECTIVE
-EINSTEIN
-ELBDPF
-ELEONORA
-ELSART
-ELSEVIER
-ENSLAPP
-ENVGRAM
-EPIPHANY
-EPSFIG
-ERNEST
-ESOMAT
-ESPART
-ESPOSITO
-EUROGAM
-EUROSPEECH
-EXOSAT
-EXOTIC
-FAIRLIE
-FARAONI
-FARGION
-FASTUS
-FECLESC
-FEDERICI
-FERMILAB
-FERRARAS
-FERRARI
-FEYNMAN
-FIELDS
-FINDER
-FISIST
-FLORIAN
-FNLVOV
-FNMILGRM
-FNRINAT
-FNUSOV
-FORMATION
-FORNENGO
-FORTRAN
-FRANCO
-FRANZKOWSKI
-FREIBURG
-FREUDENREICH
-FRITIOF
-FSHEWB
-FTLIPKIN
-FUJITA
-FUKUYAMA
-FUNAKUBO
-FURUKAWA
-FUSION
-GAILLARD
-GALLAY
-GALLEX
-GANDENBERGER
-GARBACZEWSKI
-GASPERIN
-GASPERINI
-GBRUNETTI
-GCARTER
-GEBHARD
-GELINA
-GENESIS
-GENIUS
-GENOVESE
-GENTLE
-GEOLOGY
-GEORGE
-GERALDO
-GERJIKOV
-GIACOMO
-GIAMARCHI
-GISELLA
-GISSEL
-GIUDICE
-GIUNTI
-GIVEON
-GODFREY
-GOLDBERG
-GOLDONI
-GOLDSTEI
-GONZALEZ
-GORDON
-GORSKI
-GRAESSER
-GRANAT
-GRAPESPH
-GRATUS
-GRILLI
-GRISARU
-GRNLNDDT
-GROCSE
-GROJEAN
-GROOTE
-GRUNFELD
-GUNDUC
-GUNION
-HADRON
-HAEUSLER
-HARADA
-HARVARD
-HASHIMOTO
-HAXTON
-HBCHPT
-HDECAY
-HEEMAN
-HELION
-HELIOS
-HELLAS
-HELLER
-HELMUTH
-HEMCGC
-HERMES
-HERRERA
-HERWIG
-HEWETT
-HHUANG
-HIGGSPV
-HIJING
-HILLHOUSE
-HINDAWI
-HIPPARCOS
-HIRLAM
-HIROTANI
-HITOSHI
-HKUPHYS
-HOLLIK
-HOMESTAKE
-HORWITZ
-HYLANDER
-IASSNS
-IAUNAM
-ICARUS
-ICASSP
-ICGTMP
-ICIMAF
-ICPAQGP
-ICSSUR
-IFREMER
-IFUNAM
-IMEDEA
-IMEUSP
-IMPERIAL
-IMPNWU
-INFIQC
-INFNCA
-INFNCT
-INFNFE
-INFNNA
-INFOCOM
-INTEGRAL
-INTERACTIONS
-IRAPUAN
-ISHEPP
-ISHEPPXIII
-ISHIDA
-ISOCAM
-ISOPHOT
-ISOSHIMA
-ISOSWS
-ITESRE
-ITHPII
-IVANOV
-JAKOBCZYK
-JASKOLSKI
-JETSET
-JMORRIS
-JMYANG
-JNCASR
-JPNYITP
-JUNSHIN
-JUSTINO
-KACKKACHI
-KAISER
-KAKIGI
-KALELKAR
-KALOPER
-KAMIOKANDE
-KANAZAVA
-KANAZAWA
-KARADAYI
-KARMEN
-KASCADE
-KAWAGUCHI
-KAWAMATA
-KHALIL
-KHALLER
-KHUDIAN
-KIEVSKY
-KIRTLEY
-KITADA
-KODAMA
-KOERNER
-KOHAMA
-KOHANOFF
-KOKADO
-KOMABA
-KORALW
-KOTZINIAN
-KOUNOH
-KREIMER
-KREUZER
-KSUCNR
-KULKARNI
-KUMANO
-KUPHSX
-KURIKI
-KUTSCHERA
-KYOKUGEN
-KYUSHU
-LABASTIDA
-LAFAGE
-LAGEOS
-LAMUPHYS
-LANDAU
-LARGER
-LAROCHE
-LATTICE
-LAURQCD
-LCLAVELL
-LECHEMINANT
-LEMONON
-LEONARDO
-LEPTON
-LGONZALZ
-LINERS
-LINOBS
-LONDON
-LORENTZ
-LPTENS
-LUCLUS
-LUDOVICI
-LUKIERSKI
-LYASSC
-MAARTENS
-MACHET
-MAEZONO
-MAGGIORA
-MAILBOX
-MAKEENKO
-MALTONI
-MANCINI
-MANFRED
-MANFREDINI
-MANGANO
-MARCINEK
-MARCONI
-MARKIEWIC
-MARROCCHESI
-MARTELLI
-MARTIN
-MARTIS
-MASETTI
-MASSIMO
-MATERA
-MATSUSHITA
-MATTHIAS
-MAXIMA
-MAXWELL
-MCDONALD
-MEDVEDEV
-MEGGIOLARO
-MEIERH
-MELEKHIN
-MENOTTI
-MERLIN
-MERRITT
-METDST
-MHJENSEN
-MICHAEL
-MILAGRO
-MILLENER
-MILLER
-MINERVA
-MINSEOK
-MIYAMA
-MKLIMEK
-MONTANARI
-MORICONI
-MOSAIC
-MOSFET
-MOSFETS
-MOULAY
-MOUROY
-MPIPKS
-MRENNA
-MSUHEP
-MSUPHY
-MSWORD
-MUELLER
-MUKHERJEE
-MUKOUYAMA
-MUNEHISA
-MURAKAMI
-MURAYAMA
-MURGIA
-NAGAOSA
-NAKADA
-NAKAJIMA
-NELSON
-NEUTRINO
-NICMOS
-NIKHEF
-NIRMOS
-NOJIRI
-NORDITA
-NORMAL
-NORMAN
-NUMBERS
-OBSERVATORY
-OGREID
-OHSTPY
-OKUYAMA
-OLEVIN
-OOGURI
-ORANGE
-ORDINARY
-OREGON
-ORFEUS
-ORIGINS
-ORMAND
-OSOSIRIS
-OSWALD
-OZDEMIR
-OZERNOY
-PADOVANI
-PAPERS
-PARADISE
-PARMELA
-PASCAL
-PASCHKE
-PASCOS
-PASQUALE
-PASQUINIM
-PEGASE
-PELAEZ
-PELITI
-PENDULUM
-PHCHENG
-PHENIX
-PHIDIAS
-PHOTON
-PHYSIC
-PHYSICS
-PHYSMAS
-PHYZZX
-PICTEX 
-PIERCE
-PLANCK
-PLATANIA
-POINCARE
-POLDIS
-POLONYI
-POLRAD
-POLYKARP
-POMONA
-POPOWICZ
-POSTECH
-POSTPICTEX
-POSTSCRIPT
-POTENT
-PRADISI
-PREPICTEX
-PRINCETON
-PROFILE
-PROKHOROV
-PROMETHEUS
-PROPITIUS
-PROVERO
-PSEUDOGAP
-PSTRICKS
-PUMPLIN
-PYTHIA
-QCDINS
-QFTHEP
-QUIROS
-RADCOR
-RAHIMI
-RANJITA
-RAPGAP
-RAQUEL
-README
-REGIME
-REHFELD
-RESCEU
-REVIEW
-REVTEX
-RICCARDO
-RICHERT
-RIDOLFI
-ROBERT
-ROBERTO
-ROGERIO
-ROMANINO
-ROSNER
-ROUHANI
-ROZANOV
-RUDKOVSK
-RUTGERS
-RYUTHP
-SABSNS
-SACLAY
-SACLIOGLU
-SADHAN
-SAGNOTTI
-SAKKYO
-SALASNICH
-SAMINADAYAR
-SAMPLE
-SANITA
-SAPHIR
-SARIRA
-SASAKI
-SATURNE
-SAUVAGE
-SCHAPOSNIK
-SCHNITTGER
-SCHNITZER
-SCHOLTEN
-SCHOLTZ
-SCHVELLINGER
-SCIENCE
-SCIUTO
-SCREEN
-SEMINARA
-SERGIO
-SETLUR
-SHAOJING
-SHARAPOV
-SHEIKHJABBARI
-SHELDON
-SHIFMAN
-SHIROKOV
-SHWETKETU
-SIGACT
-SIGLEX
-SIGPHON
-SIGPLAN
-SIGRAV
-SILAFAE
-SILVIO
-SIMBAD
-SIMULATOR
-SINELSHCHIKOV
-SISSAREF
-SKICAT
-SMBARR
-SMERALDI
-SMILGA
-SMUHEP
-SNBNCBS
-SNUCTP
-SOGANG
-SOUBIRAN
-SPAGNA
-SPECTRUM
-SPROCL
-SRCSSM
-STACEE
-STAMENOV
-STANEV
-STARCK
-STARNET
-STEFANO
-STIRAP
-STOILOVA
-STOJKOVIC
-STRAMAGLIA
-STRAUB
-STREETS
-STRIKMAN
-STRINGS
-STRONG
-STRUMIA
-SUBATECH
-SUCHER
-SUMMERS
-SUNYSB
-SUSLOV
-SUSSEX
-SUSYCP
-SUSYLR
-SUSYQM
-SUZUKI
-SYSTOLE
-TAKASAKI
-TAMBINI
-TATSUNO
-TAUPHY
-TAURUS
-TAVZUR
-TAYLOR
-TAYLORT
-TCDMATH
-TCHAPOVSKI
-TCILATEX
-TDASGUPTA
-TECHNION
-TEMPLE
-TEORICA
-TERAOMS
-TESCHNER
-TESFAYE
-TEVATRON
-TEXIER
-TEYSSIER
-THEMISTOCLE
-THEOCHEM
-THEORY
-THIBON
-TILLEY
-TITARCHUK
-TITCMT
-TOISFTT
-TORNQVIST
-TORRES
-TOYAMA
-TPALEV
-TRIESTE
-TRISTAN
-TRIUMF
-TROSHIN
-TSIKLAURI
-TURBOC
-TURLER
-TURNER
-UAM
-UBGVRI
-UBVRIJHK
-UBVRIJHKL
-UCDPHY
-UCRHEP
-UCSBTH
-UDEHEP
-UFSCAR
-UICHEP
-UMDEPP
-UMIACS
-UNDPDK
-UNICAMP
-UNIGRAZ
-UPCMAT
-UQMATH
-URALTSEV
-USTRON
-UTEXAS
-UVSFPFT
-VALENCIA
-VALTANCOLI
-VANKOLCK
-VASILEVICH
-VERANO
-VERBMOBIL
-VERGASSOLA
-VERITAS
-VERSION
-VESSELIN
-VICENTE
-VILLANTE
-VIRM
-VISHAL
-VISITOR
-VITALY
-VIVIANE
-VIVIANI
-VIVITRON
-VLASOV
-VNIIMS
-VOGELSANG
-VOITIKOVA
-VOLOSHIN
-WALTER
-WATPHY
-WATPHYS
-WEINBERG
-WEIZMANN
-WETDST
-WGACAT
-WITHOUT
-WUGRAV
-WUSTHOFF
-WYFFOS
-XXVIII
-XXXIII
-YAMADA
-YAMAGATA
-YAMAMOTO
-YAMAWAKI
-YELKHOVSKY
-YERPHI
-YOSHIOKA
-YOUSUF
-YUKALOV
-YUKAWA
-ZABRODIN
-ZACHOS
-ZALEWSKI
-ZFITTER
-ZHIROV
-"""
-
-KNOWN_WORDS_IN_ALL_CAPS = set(known_words_in_all_caps_str.split())
