+++ qa/qa/checks/metadata/abstract.py
@@ -3,6 +3,7 @@
 from qa.checks.base import BaseAggregateCheck
 from qa.checks.models import QaDataRegistry, OnFailurePolicy, Metadata, Result
 from qa.checks import generic
+
 # TODO: add an English language check (requires gcld3, which has no macOS arm64 wheel)
 
 
@@ -25,17 +26,29 @@ class AbstractIsValid(BaseAggregateCheck):
         return cls().run(QaDataRegistry(metadata=Metadata(abstract=abstract)))
 
     _checks = (
-        generic.NotTooShort(min_chars=5, on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
-        generic.NotTooLong(max_chars=2000, on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
+        generic.NoUtf8DecodingErrors(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="abstract"),
+        generic.NotTooShort(
+            min_chars=150,
+            on_failure_policy=OnFailurePolicy.WARN,
+            data="metadata",
+            field="abstract",
+            failure_message="Too short: must be at least 150 characters.",
+        ),
+        generic.NotTooLong(
+            max_chars=2000,
+            on_failure_policy=OnFailurePolicy.WARN,
+            data="metadata",
+            field="abstract",
+            failure_message="Too long: must be 2000 characters or fewer.",
+        ),
         generic.DoesNotBeginWithAbstract(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
         generic.NoExcessiveCapitals(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
         generic.DoesNotStartWithLowercase(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
         generic.DoesNotContainUnnecessaryEscape(
             on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"
         ),
-        generic.DoesNotContainTex(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
-        generic.DoesNotContainTexBeginEnv(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
-        generic.NoBoundaryWhitespace(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
+        generic.DoesNotContainHrefOrUrlTex(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
+        generic.DoesNotContainTexBegin(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
         generic.NoExtraWhitespace(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
         generic.NoUnnecessarySpaceInParens(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
         generic.NoHtmlElements(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
@@ -43,5 +56,4 @@ class AbstractIsValid(BaseAggregateCheck):
         generic.DoesNotContainControlCharsAllowNewlines(
             on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"
         ),
-        generic.NoUtf8DecodingErrors(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="abstract"),
     )
