+++ qa/qa/checks/metadata/journal_ref.py
@@ -30,23 +30,36 @@ class JournalRefIsValid(BaseAggregateCheck):
         return super()._run(data_registry)
 
     _checks = (
-        generic.NotTooShort(min_chars=5, on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"),
-        generic.NotTooLong(
-            max_chars=2000, on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"
+        generic.DoesNotContainPendingPublicationStatus(
+            on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="journal_ref"
         ),
-        generic.DoesNotContainUrl(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"),
-        generic.DoesNotContainDoi(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"),
-        generic.DoesNotContainBareDoi(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"),
-        generic.DoesNotContainAccepted(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"),
-        generic.DoesNotContainSubmitted(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"),
-        generic.DoesNotContainBibtex(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"),
-        generic.NoBoundaryWhitespace(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"),
-        generic.NoExtraWhitespace(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"),
+        generic.ContainsAValidYear(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="journal_ref"),
+        generic.DoesNotContainUrl(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="journal_ref"),
+        generic.DoesNotContainDoi(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="journal_ref"),
+        generic.DoesNotContainBareDoi(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="journal_ref"),
+        generic.DoesNotContainAccepted(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="journal_ref"),
+        generic.DoesNotContainSubmitted(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="journal_ref"),
+        generic.DoesNotContainBibtex(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="journal_ref"),
+        generic.NoExtraWhitespace(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="journal_ref"),
         generic.NoUnnecessarySpaceInParens(
-            on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"
+            on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="journal_ref"
         ),
         generic.DoesNotContainControlChars(
-            on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"
+            on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="journal_ref"
         ),
-        generic.NoUtf8DecodingErrors(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="journal_ref"),
+        generic.NoUtf8DecodingErrors(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="journal_ref"),
+        generic.NotTooShort(
+            min_chars=5,
+            on_failure_policy=OnFailurePolicy.WARN,
+            data="metadata",
+            field="journal_ref",
+            failure_message="Too short: must be at least 5 characters.",
+        ),
+        generic.NotTooLong(
+            max_chars=1500,
+            on_failure_policy=OnFailurePolicy.WARN,
+            data="metadata",
+            field="journal_ref",
+            failure_message="Too long: must be 1500 characters or fewer.",
+        ),
     )
