+++ qa/qa/checks/metadata/report_num.py
@@ -30,15 +30,29 @@ class ReportNumIsValid(BaseAggregateCheck):
         return super()._run(data_registry)
 
     _checks = (
-        generic.NotTooShort(min_chars=4, on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="report_num"),
-        generic.NotTooLong(max_chars=2000, on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="report_num"),
+        generic.ContainsALetterAndADigit(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="report_num"),
+        generic.NoExtraWhitespace(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="report_num"),
+        generic.NoUnnecessarySpaceInParens(
+            on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="report_num"
+        ),
+        generic.DoesNotContainControlChars(
+            on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="report_num"
+        ),
+        generic.NoUtf8DecodingErrors(on_failure_policy=OnFailurePolicy.REJECT, data="metadata", field="report_num"),
+        generic.NotTooShort(
+            min_chars=4,
+            on_failure_policy=OnFailurePolicy.WARN,
+            data="metadata",
+            field="report_num",
+            failure_message="Too short: must be at least 4 characters.",
+        ),
+        generic.NotTooLong(
+            max_chars=2000,
+            on_failure_policy=OnFailurePolicy.WARN,
+            data="metadata",
+            field="report_num",
+            failure_message="Too long: must be 2000 characters or fewer.",
+        ),
         generic.DoesNotContainUrl(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="report_num"),
         generic.DoesNotContainDoi(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="report_num"),
-        generic.ContainsLetters(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="report_num"),
-        generic.ContainsDigits(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="report_num"),
-        generic.NoBoundaryWhitespace(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="report_num"),
-        generic.NoExtraWhitespace(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="report_num"),
-        generic.NoUnnecessarySpaceInParens(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="report_num"),
-        generic.DoesNotContainControlChars(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="report_num"),
-        generic.NoUtf8DecodingErrors(on_failure_policy=OnFailurePolicy.WARN, data="metadata", field="report_num"),
     )
