+++ qa/tests/checks/metadata/test_abstract.py
@@ -76,15 +76,25 @@ class TestAbstractIsValid:
         assert result.passed
         assert not sub_result(result, "not_too_short").passed
 
-    def test_warn_begins_with_abstract(self):
+    def test_warn_begins_with_abstract_colon(self):
         result = AbstractIsValid.check("Abstract: some text")
+        assert result.passed
+        assert not sub_result(result, "does_not_begin_with_abstract").passed
+
+    def test_pass_begins_with_abstract_no_colon(self):
+        result = AbstractIsValid.check("Abstract some text")
+        assert result.passed
+        assert sub_result(result, "does_not_begin_with_abstract").passed
+
+    def test_warn_begins_with_abstract_all_caps(self):
+        result = AbstractIsValid.check("ABSTRACT: some text")
         assert result.passed
         assert not sub_result(result, "does_not_begin_with_abstract").passed
 
     def test_warn_tex_begin_no_brace(self):
         result = AbstractIsValid.check("This \\begin foo is flagged")
         assert result.passed
-        assert not sub_result(result, "does_not_contain_tex_begin_env").passed
+        assert not sub_result(result, "does_not_contain_tex_begin").passed
 
     def test_none_field_short_circuits(self):
         result = AbstractIsValid().run(QaDataRegistry(metadata=Metadata(abstract=None)))
