+++ qa/tests/checks/metadata/test_acm_class.py
@@ -14,7 +14,7 @@ def sub_result(result: Result, name: str) -> Result:
 
 class TestAcmClassIsValid:
     def test_pass_normal(self):
-        assert AcmClassIsValid.check("F.2.2; I.2.7").passed
+        assert AcmClassIsValid.check("F.2.2").passed
 
     def test_pass_none(self):
         result = AcmClassIsValid.check(None)
@@ -26,19 +26,27 @@ class TestAcmClassIsValid:
         assert result.passed
         assert result.results == []
 
+    def test_fail_semicolon_separated_list(self):
+        result = AcmClassIsValid.check("F.2.2; I.2.7")
+        assert not result.passed
+        assert not sub_result(result, "does_not_contain_semicolon").passed
+
     def test_warn_too_long(self):
-        result = AcmClassIsValid.check("x" * 1001)
+        result = AcmClassIsValid.check("x" * 161)
         assert result.passed
         assert not sub_result(result, "not_too_long").passed
 
-    def test_warn_contains_url(self):
+    def test_pass_at_max_length(self):
+        assert AcmClassIsValid.check("x" * 160).passed
+
+    def test_fail_contains_url(self):
         result = AcmClassIsValid.check("https://example.com/F.2.2")
-        assert result.passed
+        assert not result.passed
         assert not sub_result(result, "does_not_contain_url").passed
 
-    def test_warn_contains_doi(self):
+    def test_fail_contains_doi(self):
         result = AcmClassIsValid.check("doi:10.1103/F.2.2")
-        assert result.passed
+        assert not result.passed
         assert not sub_result(result, "does_not_contain_doi").passed
 
     def test_pass_space_separated(self):
