+++ qa/tests/checks/metadata/test_msc_class.py
@@ -14,7 +14,7 @@ def sub_result(result: Result, name: str) -> Result:
 
 class TestMscClassIsValid:
     def test_pass_normal(self):
-        assert MscClassIsValid.check("35K55; 65M06").passed
+        assert MscClassIsValid.check("35K55").passed
 
     def test_pass_none(self):
         result = MscClassIsValid.check(None)
@@ -27,7 +27,7 @@ class TestMscClassIsValid:
         assert result.results == []
 
     def test_warn_too_long(self):
-        result = MscClassIsValid.check("x" * 1001)
+        result = MscClassIsValid.check("x" * 161)
         assert result.passed
         assert not sub_result(result, "not_too_long").passed
 
@@ -41,10 +41,30 @@ class TestMscClassIsValid:
         assert result.passed
         assert not sub_result(result, "does_not_contain_doi").passed
 
-    def test_warn_contains_semicolon(self):
+    def test_fail_contains_semicolon(self):
         result = MscClassIsValid.check("35K55; 65M06")
-        assert result.passed
+        assert not result.passed
         assert not sub_result(result, "does_not_contain_semicolon").passed
+
+    def test_fail_extra_whitespace(self):
+        result = MscClassIsValid.check("35K55  65M06")
+        assert not result.passed
+        assert not sub_result(result, "no_extra_whitespace").passed
+
+    def test_fail_space_in_parens(self):
+        result = MscClassIsValid.check("14J60 ( Primary)")
+        assert not result.passed
+        assert not sub_result(result, "no_unnecessary_space_in_parens").passed
+
+    def test_fail_control_chars(self):
+        result = MscClassIsValid.check("35K55\t65M06")
+        assert not result.passed
+        assert not sub_result(result, "does_not_contain_control_chars").passed
+
+    def test_fail_malformed_unicode(self):
+        result = MscClassIsValid.check("35K55 \xc0\x80 65M06")
+        assert not result.passed
+        assert not sub_result(result, "no_utf8_decoding_errors").passed
 
     def test_pass_space_separated(self):
         assert MscClassIsValid.check("abc def").passed
